move graph ops init
