+++ src/development/vortex/development/__init__.py
@@ -1,5 +1,5 @@
 from .version import __version__
 
 # expose registry to top-level for shorthand
-from .exporter.utils.onnx.graph_ops import GRAPH_OPS as ONNX_GRAPH_OPS
+from .exporter.utils.onnx import GRAPH_OPS as ONNX_GRAPH_OPS
 from .networks.models import MODELS
